Move methods to show resource details to UI
